GNG-553: Refactor ContentMaintenance table into its own file
